Code/pyproject.toml
[project]
dependencies = [
  "psycopg2-binary>=2.9.9",
]
description = "Add your description here"
name = "code"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"
